backend/requirements.txt
djangorestframework
django
djangorestframework-simplejwt
django-cors-headers